+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/ConnectorInfoFactory.groovy
@@ -6,18 +6,24 @@ import com.avioconsulting.mule.testing.muleinterfaces.wrappers.connectors.HttpRe
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.connectors.SoapConsumerInfo
 
 class ConnectorInfoFactory {
-    ConnectorInfo getConnectorInfo(String fileName,
+    ConnectorInfo getConnectorInfo(Object componentLocation,
                                    String name,
-                                   String container,
-                                   Integer lineInFile,
                                    Map params) {
+        String fileName = componentLocation.fileName.get()
+        String container = componentLocation.getRootContainerName()
+        Integer lineInFile = componentLocation.lineInFile.get() as Integer
+        // the last part of our identifier should be the connector being mocked
+        // which should be a good tip as to what type of connector it is
+        def componentIdentifierOptional = componentLocation.parts.last().partIdentifier
+        assert componentIdentifierOptional.present: "Expected to find component identifier for ${name}!"
+        def identifier = componentIdentifierOptional.get().identifier
         def info = {
-            if (params['requestBuilder']?.class?.name?.endsWith('HttpRequesterRequestBuilder')) {
+            if (identifier.namespace == 'http' && identifier.name == 'request') {
                 return new HttpRequesterInfo(fileName,
                                              lineInFile,
                                              container,
                                              params)
-            } else if (params['message']?.class?.name?.endsWith('SoapMessageBuilder')) {
+            } else if (identifier.namespace == 'wsc' && identifier.name == 'consume') {
                 return new SoapConsumerInfo(fileName,
                                             lineInFile,
                                             container,