+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # 2.0.28
 * Fix issue with mocking with `raw` and media/data types
+* When figuring out the mocked connector type, use a more robust approach (component identifier) rather than hints
 
 # 2.0.27
 * Don't choke when mocking connectors that use connections